+++ packages/components/otp-input/src/styles.ts
@@ -31,6 +31,9 @@ export const otpInputVariants = tv({
         caret: "h-6",
       },
     },
+    isInvalid: {
+      true: {},
+    },
   },
   defaultVariants: { radius: "md", size: "md" },
 });