+++ packages/components/rhf/src/index.ts
@@ -3,6 +3,9 @@ import { RHFField as RHFFieldRoot } from "./rhf-field";
 import { RHFFieldError } from "./rhf-field-error";
 import { RHFInput } from "./rhf-input";
 import { RHFNumberField } from "./rhf-number-field";
+import { RHFOtpInput } from "./rhf-otp-input";
+import { RHFRadioGroup } from "./rhf-radio-group";
+import { RHFSelect } from "./rhf-select";
 import { RHFSwitch } from "./rhf-switch";
 import { RHFTextarea } from "./rhf-textarea";
 
@@ -14,4 +17,7 @@ export const RHFField = Object.assign(RHFFieldRoot, {
   Switch: RHFSwitch,
   Checkbox: RHFCheckbox,
   NumberField: RHFNumberField,
+  RadioGroup: RHFRadioGroup,
+  Select: RHFSelect,
+  OtpInput: RHFOtpInput,
 });