+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsrui/input
 
+## 0.0.17
+
+### Patch Changes
+
+- feat: rhf
+
 ## 0.0.16
 
 ### Patch Changes